fix(mcp): 재이식 정합 — rename 라우트 분리(600줄 가드)·loadServers 스코프·lint 제외
유실 브랜치 8커밋을 158 커밋 뒤처진 상태에서 cherry-pick 하며 생긴 정합 문제를 정리한다.

- `mcp.routes.ts` 가 rename 라우트 추가로 625줄 → Gate 3(600) 위반. 본문을
  `routes/mcp-server-rename.ts` 로 분리(등록만 남김) — `mcp-global-connect.ts` 와 같은 패턴.
- `loadServers` 가 `handleToggleAutoSpawn` 안에 중첩돼 스코프 밖에서 참조 불가였다
  (충돌 해결 시 HEAD 블록의 닫는 중괄호가 충돌 범위 밖에 있었던 탓). 함수 경계 복원.
- `scripts/resolve-ports.cjs` 를 eslint ignores 에 추가 — `scripts/build-info.js`·
  `scripts/eval/*.js` 와 같은 취급(CommonJS 운영 스크립트, 앱 소스 아님).
