+++ eslint.config.mjs
@@ -16,6 +16,9 @@ export default [
       "jest.config.js",
       "ecosystem.config.js",
       "scripts/build-info.js",
+      // .env 를 읽어 포트를 해석하는 부팅 전 스크립트 — pm2/next 설정이 require 로 로드한다.
+      // build-info.js 와 동일 취급(앱 소스 아님).
+      "scripts/resolve-ports.cjs",
       // CommonJS 로 dist 를 직접 로드하는 운영 스크립트(빌드 산출물 require 필요) —
       // build-info.js 와 동일 취급. 앱 소스가 아니라 lint 대상에서 제외한다.
       "scripts/eval/*.js",
